now autograded files that timeout actually get an xml file
the file is made using the new .py file, `make_timeout_xml.py`.  messages are also printed to the screen, and written into the .out files that it was interrupted.

a potential weakness: i'm using `SIGINT` instead of the default, so that pytest will generate the messages at the end if its run.  It's possible this isn't strong enough a signal, but it worked for me just now, so I'm going for it.